# testing-environment-backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/projects/{projectId}/test-suites": {
      "post": {
        "operationId": "TestSuitesController_create",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTestSuiteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Test suites"
        ]
      },
      "get": {
        "operationId": "TestSuitesController_list",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTestSuitesResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Test suites"
        ]
      }
    },
    "/projects/{projectId}/test-suites/compile-flow": {
      "post": {
        "operationId": "TestSuitesController_compileFlow",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompileFlowDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Test suites"
        ]
      }
    },
    "/projects/{projectId}/test-suites/import/preview": {
      "post": {
        "operationId": "TestSuitesController_previewImport",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportPreviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Test suites"
        ]
      }
    },
    "/projects/{projectId}/test-suites/import/generate-flow": {
      "post": {
        "operationId": "TestSuitesController_generateImportedFlow",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateImportedFlowDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Test suites"
        ]
      }
    },
    "/projects/{projectId}/test-suites/{suiteId}/revisions": {
      "get": {
        "operationId": "TestSuitesController_revisions",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "suiteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Test suites"
        ]
      }
    },
    "/projects/{projectId}/test-suites/{suiteId}/revisions/compare": {
      "get": {
        "operationId": "TestSuitesController_compareRevisions",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "suiteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Test suites"
        ]
      }
    },
    "/projects/{projectId}/test-suites/{suiteId}/revisions/{revisionId}/publish": {
      "post": {
        "operationId": "TestSuitesController_publishRevision",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "suiteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "revisionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Test suites"
        ]
      }
    },
    "/projects/{projectId}/test-suites/{suiteId}": {
      "get": {
        "operationId": "TestSuitesController_find",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "suiteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestSuiteResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Test suites"
        ]
      },
      "patch": {
        "operationId": "TestSuitesController_update",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "suiteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTestSuiteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Test suites"
        ]
      },
      "delete": {
        "operationId": "TestSuitesController_delete",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "suiteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Test suites"
        ]
      }
    },
...
  },
  "info": {
    "title": "Backend Test Runner API",
    "description": "MVP backend control plane and local runner engine.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ]
      },
...
      "CreateTestSuiteDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "yamlContent": {
            "type": "string"
          },
          "sourceMode": {
            "type": "string",
            "enum": [
              "VISUAL",
              "RAW_YAML"
            ]
          },
          "visualFlow": {
            "type": "object"
          }
        },
        "required": [
          "name",
          "yamlContent"
        ]
      },
      "TestSuiteResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "yamlContent": {
            "type": "string"
          },
          "visualFlow": {
            "type": "object"
          },
          "currentRevision": {
            "type": "object"
          },
          "publishedRevision": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "projectId",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedTestSuitesResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TestSuiteResponseDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "CompileFlowDto": {
        "type": "object",
        "properties": {
          "visualFlow": {
            "type": "object"
          }
        },
        "required": [
          "visualFlow"
        ]
      },
      "ImportPreviewDto": {
        "type": "object",
        "properties": {
          "sourceType": {
            "type": "string",
            "enum": [
              "OPENAPI",
              "POSTMAN",
              "BRUNO",
              "CURL",
              "MANUAL"
            ]
          },
          "content": {
            "type": "string"
          },
          "files": {
            "type": "object"
          },
          "manualRequest": {
            "type": "object"
          }
        },
        "required": [
          "sourceType"
        ]
      },
      "GenerateImportedFlowDto": {
        "type": "object",
        "properties": {
          "suiteName": {
            "type": "string"
          },
          "template": {
            "type": "string",
            "enum": [
              "SMOKE_TEST",
              "AUTHENTICATED_JOURNEY",
              "CRUD_LIFECYCLE",
              "ASYNC_POLLING",
              "READINESS_TEST"
            ]
          },
          "operations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "acknowledgeDestructive": {
            "type": "boolean"
          },
          "options": {
            "type": "object"
          }
        },
        "required": [
          "suiteName",
          "template",
          "operations"
        ]
      },
      "UpdateTestSuiteDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "yamlContent": {
            "type": "string"
          },
          "sourceMode": {
            "type": "string",
            "enum": [
              "VISUAL",
              "RAW_YAML"
            ]
          },
          "visualFlow": {
            "type": "object"
          }
        }
      },
...
    }
  }
}